Moved unit tests for VcfVariantsReader to NucleotideSampleData
